// src/db/Model/Employee.ts
export default interface IEmployee {
    empID: string;
    name: string;
    passwordHash: string;
    userName: string;
    address: string;
    nic: string;
    userRole: string;
    joinedDate: string;
    isAvailable: boolean;
    jobTitle: string;
    birthDay: string;
    orgID: string;
    contactNum: number;
    email: string;
    image: string;
}